requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
python-dateutil>=2.8.2 